Fixed interfaces.
