+++ Code/Engine/Core/Module.h
@@ -6,9 +6,15 @@
 
 #pragma once
 
+#include <memory>
+
 
 #ifndef CORE_EXPORTS
 #define CORE_API __declspec (dllimport)
 #else
 #define CORE_API __declspec (dllexport)
 #endif
+
+
+/** Physics world */
+typedef std::shared_ptr<class IPhysicsWorld> TPhysicsWorldPtr;
