+++ Code/Engine/GUI/Widgets.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include "Module.h"
-#include "World/Visual.h"
+#include "GUI/WidgetInterface.h"
 
 #include <string>
 #include <functional>
@@ -18,25 +18,22 @@ struct LcWidgetData
 {
     LcVector3 pos;	// [0,0] - leftBottom, x - right, y - up
     LcSizef size;	// widget size in pixels
-	bool visible;
-	bool active;
-	//
-	LcWidgetData() : pos(LcDefaults::ZeroVec3), size(LcDefaults::ZeroSize), visible(true), active(false) {}
-	//
-	LcWidgetData(LcVector3 inPos, LcSizef inSize, bool inVisible = true, bool inActive = false)
-		: pos(inPos), size(inSize), visible(inVisible), active(inActive)
-	{
-	}
+    bool visible;
+    bool disabled;
     //
-    LcWidgetData(LcVector2 inPos, LcSizef inSize, bool inVisible = true, bool inActive = false)
-        : pos(To3(inPos)), size(inSize), visible(inVisible), active(inActive)
+    LcWidgetData() : pos(LcDefaults::ZeroVec3), size(LcDefaults::ZeroSize), visible(true), disabled(false) {}
+    //
+    LcWidgetData(LcVector3 inPos, LcSizef inSize, bool inVisible = true, bool inDisabled = false)
+        : pos(inPos), size(inSize), visible(inVisible), disabled(inDisabled)
+    {
+    }
+    //
+    LcWidgetData(LcVector2 inPos, LcSizef inSize, bool inVisible = true, bool inDisabled = false)
+        : pos(To3(inPos)), size(inSize), visible(inVisible), disabled(inDisabled)
     {
     }
 };
 
-
-/** Widget texture component */
-typedef struct LcSpriteTextureComponent LcWidgetTextureComponent;
 
 /**
 * Widget text component */
@@ -64,17 +61,6 @@ struct LcWidgetTextComponent : public IVisualComponent
 };
 
 
-/** Button state */
-enum class EBtnState : int
-{
-    Idle,
-    Over,
-    Pressed
-};
-
-/** Button click handler */
-typedef std::function<void()> FBtnHandler;
-
 /**
 * Widget button component */
 struct GUI_API LcWidgetButtonComponent : public IVisualComponent
@@ -101,15 +87,6 @@ public:// IVisualComponent interface implementation
 
 };
 
-
-/** Checkbox state */
-enum class ECheckboxState : int
-{
-    Unchecked,
-    UncheckedHovered,
-    Checked,
-    CheckedHovered
-};
 
 /**
 * Widget checkbox component */
@@ -141,9 +118,6 @@ public:// IVisualComponent interface implementation
 };
 
 
-/** Widget click handler */
-typedef std::function<void()> LcClickHandler;
-
 /**
 * Widget click component */
 struct LcWidgetClickComponent : public IVisualComponent
@@ -161,9 +135,6 @@ struct LcWidgetClickComponent : public IVisualComponent
 };
 
 
-/** Widget check handler */
-typedef std::function<void(bool)> LcCheckHandler;
-
 /**
 * Widget check component */
 struct LcWidgetCheckComponent : public IVisualComponent
@@ -177,98 +148,6 @@ struct LcWidgetCheckComponent : public IVisualComponent
     LcWidgetCheckComponent(LcCheckHandler inHandler) : handler(inHandler) {}
     // IVisualComponent interface implementation
     virtual EVCType GetType() const override { return EVCType::CheckHandler; }
-
-};
-
-
-/**
-* Widget interface */
-class IWidget : public IVisualBase
-{
-public:
-    void AddTextComponent(const std::wstring& inText, LcColor4 inTextColor = LcDefaults::Black4,
-        const std::wstring& inFontName = L"Calibri", unsigned short inFontSize = 22, LcFontWeight inFontWeight = LcFontWeight::Normal)
-    {
-        AddComponent(std::make_shared<LcWidgetTextComponent>(inText, inFontName, inFontSize, inTextColor, inFontWeight));
-    }
-    //
-    void AddTextureComponent(const std::string& texture)
-    {
-        AddComponent(std::make_shared<LcVisualTextureComponent>(texture));
-    }
-    //
-    void AddTextureComponent(const LcBytes& inData)
-    {
-        AddComponent(std::make_shared<LcVisualTextureComponent>(inData));
-    }
-    void AddButtonComponent()
-    {
-        AddComponent(std::make_shared<LcWidgetButtonComponent>());
-    }
-    void AddButtonComponent(const std::string& texture, LcVector2 idlePos, LcVector2 overPos, LcVector2 pressedPos)
-    {
-        AddComponent(std::make_shared<LcVisualTextureComponent>(texture));
-        AddComponent(std::make_shared<LcWidgetButtonComponent>(idlePos, overPos, pressedPos));
-    }
-    void AddCheckboxComponent(const std::string& texture, LcVector2 uncheckedPos, LcVector2 uncheckedHoveredPos,
-        LcVector2 checkedPos, LcVector2 checkedHoveredPos)
-    {
-        AddComponent(std::make_shared<LcVisualTextureComponent>(texture));
-        AddComponent(std::make_shared<LcWidgetCheckboxComponent>(uncheckedPos, uncheckedHoveredPos, checkedPos, checkedHoveredPos));
-    }
-    void AddClickHandlerComponent(LcClickHandler handler, bool addDefaultSkin = true)
-    {
-        AddComponent(std::make_shared<LcWidgetClickComponent>(handler));
-
-        if (addDefaultSkin) AddButtonComponent("../../Assets/button.png",
-            LcVector2(2.0f, 2.0f), LcVector2(2.0f, 44.0f), LcVector2(2.0f, 86.0f));
-    }
-    void AddCheckHandlerComponent(LcCheckHandler handler, bool addDefaultSkin = true)
-    {
-        AddComponent(std::make_shared<LcWidgetCheckComponent>(handler));
-
-        if (addDefaultSkin) AddCheckboxComponent("../../Assets/checkbox.png",
-            LcVector2(0.0f, 0.0f), LcVector2(32.0f, 0.0f), LcVector2(0.0f, 32.0f), LcVector2(32.0f, 32.0f));
-    }
-    //
-    LcWidgetTextComponent* GetTextComponent() const { return (LcWidgetTextComponent*)GetComponent(EVCType::Text).get(); }
-    //
-    LcVisualTextureComponent* GetTextureComponent() const { return (LcVisualTextureComponent*)GetComponent(EVCType::Texture).get(); }
-    //
-    LcWidgetButtonComponent* GetButtonComponent() const { return (LcWidgetButtonComponent*)GetComponent(EVCType::Button).get(); }
-    //
-    LcWidgetCheckboxComponent* GetCheckboxComponent() const { return (LcWidgetCheckboxComponent*)GetComponent(EVCType::Checkbox).get(); }
-    //
-    LcWidgetClickComponent* GetClickHandlerComponent() const { return (LcWidgetClickComponent*)GetComponent(EVCType::ClickHandler).get(); }
-    //
-    LcWidgetCheckComponent* GetCheckHandlerComponent() const { return (LcWidgetCheckComponent*)GetComponent(EVCType::CheckHandler).get(); }
-
-
-public:
-    /**
-    * Keyboard key event */
-    virtual void OnKeyboard(int btn, LcKeyState state) = 0;
-    /**
-    * Recreate font after World scale change */
-    virtual void RecreateFont() = 0;
-	/**
-	* Set active state */
-	virtual void SetActive(bool active) = 0;
-	/**
-	* Active state */
-	virtual bool IsActive() const = 0;
-    /**
-    * Set mouse over */
-    virtual void SetHovered(bool active) = 0;
-    /**
-    * Is mouse over */
-    virtual bool IsHovered() const = 0;
-	/**
-	* Set focused state */
-	virtual void SetFocus(bool focus) = 0;
-	/**
-	* Focused state */
-	virtual bool HasFocus() const = 0;
 
 };
 
@@ -295,9 +174,9 @@ public:// IWidget interface implementation
     //
     virtual void RecreateFont() override {}
     //
-    virtual void SetActive(bool inActive) override { widget.active = inActive; }
+    virtual void SetDisabled(bool inDisabled) override { widget.disabled = inDisabled; }
     //
-    virtual bool IsActive() const override { return widget.active; }
+    virtual bool IsDisabled() const override { return widget.disabled; }
     //
     virtual void SetHovered(bool inHovered) override { hovered = inHovered; }
     //
