+++ Code/Samples/Windows/ApplicationSample/ApplicationSample.cpp
@@ -7,12 +7,11 @@
 #include "framework.h"
 #include "ApplicationSample.h"
 #include "Application/AppConfig.h"
-#include "Application/Application.h"
 #include "Application/Windows/Module.h"
 #include "RenderSystem/RenderSystemDX10/Module.h"
 #include "World/WorldInterface.h"
+#include "GUI/WidgetInterface.h"
 #include "World/Sprites.h"
-#include "GUI/Widgets.h"
 #include "Core/LCUtils.h"
