+++ Code/Samples/Windows/WidgetSample/WidgetSample.cpp
@@ -6,13 +6,11 @@
 
 #include "framework.h"
 #include "WidgetSample.h"
-#include "Application/AppConfig.h"
-#include "Application/Application.h"
 #include "Application/Windows/Module.h"
 #include "RenderSystem/RenderSystemDX10/Module.h"
+#include "World/SpriteInterface.h"
 #include "World/WorldInterface.h"
-#include "World/Sprites.h"
-#include "GUI/Widgets.h"
+#include "GUI/WidgetInterface.h"
 #include "Core/LCUtils.h"
