+++ Projects/Windows/Application/Application.vcxproj.filters
@@ -27,14 +27,14 @@
     <ClInclude Include="pch.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Code\Engine\Application\Application.h">
-      <Filter>Header Files\Application</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\..\Code\Engine\Application\Module.h">
       <Filter>Header Files\Application</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\Code\Engine\Application\AppConfig.h">
       <Filter>Source Files\Application</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\Code\Engine\Application\ApplicationInterface.h">
+      <Filter>Header Files\Application</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
